Correct sync_to_model_test
